Created TableBuilder for BW engine
