+++ Source/QED/BreitWheelerEngineTableBuilder.cpp
@@ -1,6 +1,5 @@
 #include "BreitWheelerEngineTableBuilder.H"
 
-
 //Include the full BW engine with table generation support
 //(after some consistency tests)
 #ifdef PXRMP_CORE_ONLY
@@ -12,3 +11,35 @@
 #endif
 #include "breit_wheeler_engine.hpp"
 //_______________________________________________
+
+
+using PicsarBreitWheelerEngine = picsar::multi_physics::
+    breit_wheeler_engine<amrex::Real, QedUtils::DummyStruct>;
+
+using PicsarBreitWheelerCtrl =
+    picsar::multi_physics::breit_wheeler_engine_ctrl<amrex::Real>;
+
+
+void
+BreitWheelerEngineTableBuilder::compute_table
+    (PicsarBreitWheelerCtrl ctrl,
+     BreitWheelerEngineInnards& innards) const
+{
+    PicsarBreitWheelerEngine bw_engine(
+        std::move(QedUtils::DummyStruct()), 1.0, ctrl);
+
+    bw_engine.compute_dN_dt_lookup_table();
+    bw_engine.compute_cumulative_pair_table();
+
+    auto bw_innards_picsar = bw_engine.export_innards();
+
+    //Copy data in a GPU-friendly data-structure
+    innards.ctrl = bw_innards_picsar.bw_ctrl;
+    innards.TTfunc_coords.assign(bw_innards_picsar.TTfunc_table_coords_ptr,
+        bw_innards_picsar.TTfunc_table_coords_ptr +
+        bw_innards_picsar.TTfunc_table_coords_how_many);
+    innards.TTfunc_data.assign(bw_innards_picsar.TTfunc_table_data_ptr,
+        bw_innards_picsar.TTfunc_table_data_ptr +
+        bw_innards_picsar.TTfunc_table_data_how_many);
+    //____
+}
