+++ Source/QED/BreitWheelerEngineWrapper.cpp
@@ -184,24 +184,7 @@ std::vector<char> export_lookup_tables_data () const
 void BreitWheelerEngine::computes_lookup_tables (
     PicsarBreitWheelerCtrl ctrl)
 {
-    //Lambda is not actually used if S.I. units are enabled
-    PicsarBreitWheelerEngine bw_engine(
-        std::move(QedUtils::DummyStruct()), 1.0, ctrl);
-
-    bw_engine.compute_dN_dt_lookup_table();
-    bw_engine.compute_cumulative_pair_table();
-
-    auto bw_innards_picsar = bw_engine.export_innards();
-
-    //Copy data in a GPU-friendly data-structure
-    innards.ctrl = bw_innards_picsar.bw_ctrl;
-    innards.TTfunc_coords.assign(bw_innards_picsar.TTfunc_table_coords_ptr,
-        bw_innards_picsar.TTfunc_table_coords_ptr +
-        bw_innards_picsar.TTfunc_table_coords_how_many);
-    innards.TTfunc_data.assign(bw_innards_picsar.TTfunc_table_data_ptr,
-        bw_innards_picsar.TTfunc_table_data_ptr +
-        bw_innards_picsar.TTfunc_table_data_how_many);
-    //____
+    table_builder.compute_table(ctrl, innards);
 }
 
 #endif
