+++ Source/QED/QedWrapperCommons.H
@@ -16,27 +16,4 @@ namespace QedUtils{
     struct DummyStruct{};
 };
 
-//Control structures used by the engines
-struct PicsarBreitWheelerCtrl
-{
-              _REAL chi_phot_min =
-            static_cast<_REAL>(__breit_wheeler_min_chi_phot);
-
-          _REAL chi_phot_tdndt_min =
-            static_cast<_REAL>(__breit_wheeler_min_tdndt_chi_phot);
-          _REAL chi_phot_tdndt_max =
-            static_cast<_REAL>(__breit_wheeler_max_tdndt_chi_phot);
-          size_t chi_phot_tdndt_how_many =
-            __breit_wheeler_how_many_tdndt_chi_phot;
-
-          _REAL chi_phot_tpair_min =
-            static_cast<_REAL>(__breit_wheeler_min_tpair_chi_phot);
-          _REAL chi_phot_tpair_max =
-            static_cast<_REAL>(__breit_wheeler_max_tpair_chi_phot);
-          size_t chi_phot_tpair_how_many =
-            __breit_wheeler_how_many_tpair_chi_phot;
-          size_t chi_frac_tpair_how_many =
-            __breit_wheeler_chi_frac_tpair_how_many;
-};
-
 #endif //WARPX_amrex_qed_wrapper_commons_h_
